Consolidate identify_compression and from_file* methods
Extends the implementation of custom compression magic in the previous
commit by moving all compression identification logic into the
`CompressionManager.identify_compression` method and updating the
existing barman code to use it.

The main change to the existing barman code is updating
`WalFileInfo.from_file` so that it accepts a `CompressionManager`.
`from_file` is then able to call `CompressionManager.identify_compression`
and use the value of `unidentified_compression` if nothing can be found.
It is important that `unidentified_compression` remains a separate option
even though it can be obtained from the `CompressionManager` because
we need to be able to omit it when calling the function in the context
of wal-archival, where the inability to identify compression means that
the WAL has not been compressed.

The `identify_compression` function itself now has a special case for
`file_type == "custom"` which will instantiate the compressor and call
`validate` on the instance. All other compressors are validated using
the class method as before.

Because `WalFileInfo.from_file` is now capable of setting the
compression magic for custom compression and the built-in compressors,
`from_file_with_compression` is no longer needed and therefore removed.

The changes are unfortunately quite disruptive to the tests, mainly
because many tests now need to either mock or create a
`CompressionManager` so that `WalFileInfo.from_file` can run.

The following specific changes are made to the tests:

- An additional test is added to `TestCompressionManager` which checks
  we can still create a default custom compressor when no
  `custom_compression_magic` is set.
- The test for the standalone `identify_compression` function is removed
  because the function itself is also removed.
- The `TestCommandCompressors` tests are updated to use a
  `CompressionManager` instead of the now-removed standalone
  `identify_compression` function.
- The `TestWalFileInfo` tests are updated to use a mock `CompressionManager`.
- `test_xlog.Test.test_decode_history_file` is updated to use a
  `CompressionManager`.

Finally, `TestFileWalArchiver.test_archive_wal` is updated to use a
real `CompressionManager` instead of a mock. In order for this to work
we have to patch the compression registry so that the first compressor
matching the gzip magic is a native python compressor and not an
external command.

sfaafs